+++ lib/chef/providers.rb
@@ -36,7 +36,6 @@ require_relative "provider/mount"
 require_relative "provider/noop"
 require_relative "provider/package"
 require_relative "provider/powershell_script"
-require_relative "provider/osx_profile"
 require_relative "provider/remote_directory"
 require_relative "provider/remote_file"
 require_relative "provider/route"
